requirements.txt
dj-static==0.0.6
Django==2.2.9
pytz==2019.3
static3==0.7.0
gunicorn==20.0.4
django-allauth==0.42.0
django-multiselectfield==0.1.12
django-bootstrap3==12.0.3
Pillow==6.2.2
django-encrypted-model-fields==0.5.8
django-phonenumber-field==5.0.0
